api_gateway/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
SQLAlchemy
pika
requests
PyJWT